Move logic for #encrypt & #decrypt to a Cipher class
